Multizsite53 - #2 (#1810)
* Converts S53 to One File, cleans up old files.

Converted S53 into one Multi-Z file instead of five different Z levels.
Moved 106 PD to z1_admin. Got rid of old files that are no longer used or were never used but were being stored in the map folder.

Did you know 106's PD had its own Z level with Nothing Else on it, and it took up roughly 33x20 tiles worth of that 255x255?

The CI base is already functional on the admin-z, so the file is either a duplicate or not used at all.

-Removed admin_levels from define (handeled in a separate file, not the main map file)
-Removed empty_levels from define (unused entirely)
-Removed accessible_z_levels from define. (You should not be able to access any of the outer map edges irregardless due to mapping design, but there is no reason we should have percentile chances for drifting onto other Z-levels, we're not in space.)
-Added sealed_levels define for all player-z's which prevents players from stepping onto the edge of the map even if they did manage to somehow navigate past unbreakable rock.